refactor(bpe): split the tables file, document the packed merge value once
